address sophies comments
