+++ kafka-client/src/test/java/dev/responsive/utils/SubPartitionerTest.java
@@ -18,14 +18,21 @@ package dev.responsive.utils;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
 
 import java.util.List;
+import java.util.OptionalInt;
 import java.util.stream.Collectors;
+import org.apache.kafka.common.config.ConfigException;
 import org.apache.kafka.common.utils.Bytes;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 class SubPartitionerTest {
+
+  private static final TableName NAME = new TableName("table");
+  private static final String CHANGELOG_TOPIC_NAME = "changelog";
 
   @Test
   public void shouldMapPartitionsToLargerSpace() {
@@ -64,6 +71,72 @@ class SubPartitionerTest {
         SubPartitioner.SALT,
         is(31)
     );
+  }
+
+  @Test
+  public void shouldConfigureSubPartitionerWhenDesireIsNotMultipleOfActual() {
+    // Given:
+    final var actualRemoteCount = OptionalInt.of(128);
+    final int kafkaPartitions = 32;
+    final int desiredPartitions = 129;
+
+    // When:
+    final SubPartitioner subPartitioner = SubPartitioner.create(
+        actualRemoteCount,
+        kafkaPartitions,
+        desiredPartitions,
+        NAME,
+        CHANGELOG_TOPIC_NAME
+    );
+
+    // Then:
+    assertThat(subPartitioner.getFactor(), is(4));
+  }
+
+  @Test
+  public void shouldConfigureSubPartitionerWhenDesiredIsNegativeOne() {
+    // Given:
+    final var actualRemoteCount = OptionalInt.of(32);
+    final int kafkaPartitions = 32;
+    final int desiredPartitions = -1;
+
+    // When:
+    final SubPartitioner subPartitioner = SubPartitioner.create(
+        actualRemoteCount,
+        kafkaPartitions,
+        desiredPartitions,
+        NAME,
+        CHANGELOG_TOPIC_NAME
+    );
+
+    // Then:
+    assertThat(subPartitioner.getFactor(), is(1));
+  }
+
+  @Test
+  public void shouldThrowExceptionIfActualPartitionsDoesNotMatchComputedPartitions() {
+    // Given:
+    final var actualRemoteCount = OptionalInt.of(129);
+    final int kafkaPartitions = 32;
+    final int desiredPartitions = 129;
+
+    // Expect:
+    final ConfigException error = Assertions.assertThrows(
+        ConfigException.class,
+        () -> SubPartitioner.create(
+            actualRemoteCount,
+            kafkaPartitions,
+            desiredPartitions,
+            NAME,
+            CHANGELOG_TOPIC_NAME
+        )
+    );
+
+    // Then:
+    assertThat(
+        error.getMessage(),
+        containsString("was configured to 129, which given 32 partitions in "
+            + "kafka topic changelog would result in 128 remote partitions"));
   }
 
 }
